pyproject.toml
[project]
name = "test-optimization-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "langchain>=1.2.14",
    "langchain-cerebras>=0.8.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
